removed gather_acc.hpp
