+++ domain/include/cstone/primitives/primitives_acc.hpp
@@ -8,7 +8,7 @@
  */
 
 /*! @file
- * @brief  CPU/GPU Wrapper of basic algorithms
+ * @brief  CPU/GPU wrappers for basic algorithms
  *
  * @author Sebastian Keller <sebastian.f.keller@gmail.com>
  */
@@ -18,6 +18,7 @@
 #include <algorithm>
 #include <type_traits>
 
+#include "gather.hpp"
 #include "cstone/primitives/primitives_gpu.h"
 
 namespace cstone
@@ -42,6 +43,20 @@ void fill(T* first, T* last, T value)
 
     if constexpr (useGpu) { fillGpu(first, last, value); }
     else { std::fill(first, last, value); }
+}
+
+template<bool useGpu, class IndexType, class ValueType>
+void gatherAcc(std::span<const IndexType> ordering, const ValueType* source, ValueType* destination)
+{
+    if constexpr (useGpu) { gatherGpu(ordering.data(), ordering.size(), source, destination); }
+    else { gather(ordering, source, destination); }
+}
+
+template<bool useGpu, class IndexType, class ValueType>
+void scatterAcc(std::span<const IndexType> ordering, const ValueType* source, ValueType* destination)
+{
+    if constexpr (useGpu) { scatterGpu(ordering.data(), ordering.size(), source, destination); }
+    else { scatter(ordering, source, destination); }
 }
 
 } // namespace cstone
